Refactored message handling to fully comply with WS spec.
Messages were moved from either a char* or ByteArray to a discrete SocketMessage object, which is a struct containing data and type. This change allows this WebSocket server implementation to now pass 100% of the Autobahn Testsuite WebSocket tests.
This change also moves the different handling of binary or string data into the end user application.